coin points added

+++ Assets/scripts/GameManager.cs
@@ -22,7 +22,7 @@ public class GameManager : MonoBehaviour
             float distance = Vector3.Distance(player.position,startingPos);
             //Debug.Log("Counter modulus 5: "+ counter%5);
             //Debug.Log("playerPosition: "+player.transform.position);
-            if(distance%distanceToChangeLevelAt== 0){
+            if((int)distance%distanceToChangeLevelAt== 0){
                 Instantiate(checkPoint,player.transform.position+player.transform.forward * offset,Quaternion.identity);
             }
         yield return new WaitForSeconds(1);
